Created TestHand class to be used to drive simulated hands in play mode tests.
