+++ Assets/MixedRealityToolkit.Tests/PlayModeTests/TestHand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2a7da2f9d3caf54eba05b2ef7664c51
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
